feat: observatory setController
