+++ cli/src/commands/server.ts
@@ -1,8 +1,10 @@
 import {assertNonNullish, isNullish} from '@dfinity/utils';
 import type {OutgoingHttpHeaders} from 'http';
 import {createServer, type IncomingMessage, type Server, type ServerResponse} from 'node:http';
-import {setController} from '../services/console.services';
+import {consoleModule} from '../modules/console';
+import {observatory} from '../modules/observatory';
 import {buildContext} from '../services/context.services';
+import {setController} from '../services/controller.services';
 import {transfer} from '../services/ledger.services';
 import type {CliContext} from '../types/context';
 import {nextArg} from '../utils/args.utils';
@@ -61,10 +63,14 @@ const buildServer = ({
     // If the CLI was build for the satellite but the /console/ is queried, then the feature is not supported.
     const consoleBuild = process.env.CLI_BUILD === 'console';
 
-    if (consoleBuild && command === 'console') {
+    if (consoleBuild && ['console', 'observatory'].includes(command)) {
       switch (subCommand) {
         case 'controller':
-          await setController({context, searchParams});
+          await setController({
+            context,
+            searchParams,
+            key: command === 'observatory' ? observatory.key : consoleModule.key
+          });
           done();
           return;
       }
