+++ cli/src/modules/observatory.ts
@@ -3,10 +3,12 @@ import {DEV_OBSERVATORY} from '../constants/constants';
 import {Module} from '../services/modules.services';
 import type {ModuleDescription} from '../types/module';
 
+export const OBSERVATORY_CANISTER_ID = 'klbfr-lqaaa-aaaak-qbwsa-cai';
+
 const OBSERVATORY: ModuleDescription = {
   key: 'observatory',
   name: 'Observatory',
-  canisterId: 'klbfr-lqaaa-aaaak-qbwsa-cai'
+  canisterId: OBSERVATORY_CANISTER_ID
 };
 
 export const initObservatoryModule = (): Module =>
