+++ cli/src/services/console.services.ts
@@ -1,5 +1,3 @@
-import {ICManagementCanister} from '@dfinity/ic-management';
-import {Principal} from '@dfinity/principal';
 import {assertNonNullish, fromNullable, toNullable, uint8ArrayToHexString} from '@dfinity/utils';
 import {commitProposal, initProposal, submitProposal, uploadAsset} from '@junobuild/console';
 import type {ENCODING_TYPE} from '@junobuild/storage';
@@ -8,50 +6,6 @@ import {CONSOLE_CANISTER_ID} from '../modules/console';
 import {buildConsoleParams} from '../modules/satellite/console.config';
 import type {CliContext} from '../types/context';
 import {loadWasm} from '../utils/wasm.utils';
-import {getConsoleActor} from './actor.services';
-
-export const setController = async ({
-  context,
-  searchParams
-}: {
-  context: CliContext;
-  searchParams: URLSearchParams;
-}) => {
-  const {agent} = context;
-
-  const id = searchParams.get('id') ?? '';
-
-  // Set the controller to the canister.
-  const {updateSettings, canisterStatus} = ICManagementCanister.create({
-    agent
-  });
-
-  const {
-    settings: {controllers}
-  } = await canisterStatus(Principal.from(CONSOLE_CANISTER_ID));
-
-  await updateSettings({
-    canisterId: Principal.from(CONSOLE_CANISTER_ID),
-    settings: {
-      controllers: [...controllers.map((p) => p.toText()), id]
-    }
-  });
-
-  // Add controller to the memory of the canister to allow guarded calls.
-  const {set_controllers} = await getConsoleActor({
-    agent,
-    canisterId: CONSOLE_CANISTER_ID
-  });
-
-  await set_controllers({
-    controllers: [Principal.fromText(id)],
-    controller: {
-      metadata: [],
-      scope: {Admin: null},
-      expires_at: []
-    }
-  });
-};
 
 export const installRelease = async ({
   context,
